refactor(termiod): move the session-protocol codec into TermioShared
The framing, the control payloads, and the encode/decode tables are the
protocol itself, not the Mac's way of reaching a daemon. They move to
Shared/Sources/TermioShared/TermiodProtocol.swift for the reason
WireProtocol.swift already lives there: the phone will attach over the same
codec instead of a second Swift copy that drifts silently (RFC
20260824-ios-as-device-client.md §D5).

Everything route-shaped stays in TermiodClient.swift, which now extends the
same namespace: the Unix socket, the ssh-stdio child, `ssh -G` resolution,
daemon spawn and discovery, the device registry, and the handshake's own
bookkeeping. `performHello` builds its `hello` through the shared
`helloPayload(role:caps:client:)` and passes this app's client banner, so the
codec never claims to be a particular client.

No wire change: same frames, same JSON, same field names.
